Changed Cog Allocations Added subcommands to all commands with the 'actions' argument.
